Remove the need for git submodules
Git submodules are overly complex.
I'd prefer to treat the cowfiles like any other dependency and fetch
at build time not at clone time.
This change also simplifies cleaning cowfiles between builds which
was previously a manual task.
It is also more academically pure in that we are not writing "generated"
content to the source directory which then has to be include in .gitignore.